+++ changelog.md
@@ -9,6 +9,7 @@ improvements may be omitted.)_
 __[maintenance]__
 
 * Update certificate information for Blender LTS installer.
+* Update certificate information for KeePass installers.
 * NLog library is updated from 5.3.4 to 5.4.0.
 
 ## Version 2025.01.31.0